refactor(lib): extract applyMask dispatch into mask handler registry
